+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+V1.2 - 5/11/2025
+Change an error message to say the right thing if the pack wasn't found and added more packs.
+- 1st Birthday Skin Pack
+- 2nd Birthday Skin Pack
+- 3rd Birthday Skin Pack
+- MINECON Earth 2017 Skin Pack
+- 4th Birthday Skin Pack
+- Skin Pack 1 - Classic
+- Skin Pack 2 - Classic
+- Magic Spells
+- Dinosaurs Add-On 2.1
+- More TNT! Add-On
+- DWELLERS 2.0 Add-On
+- Cherry Tree Temple
+- Added github logo on top left to get back to repository
+- 
+
 V1.1 - 5/7/2025
 Added more packs
 - PureBDcraft HD UI